Time: 8 ms (35.79%), Space: 57.9 MB (64.64%) - LeetHub

+++ 1631-number-of-sub-arrays-with-odd-sum/1631-number-of-sub-arrays-with-odd-sum.java
@@ -16,7 +16,7 @@ class Solution {
             }
             count = count % 1000000007;
         }
-        return (int)count;
+        return count;
 
     }
 }
